refactor(backend): centralize iRacing YAML preprocessing in pitwall-local
